+++ NEWS.md
@@ -1,3 +1,8 @@
+# scorecard 0.3.9
+
+* fixed a bug in woebin function.
+* the var_filter function supports to filter variable via step and vif.
+
 # scorecard 0.3.8
 
 * fixed a bug in woebin function when breaks_list is provided.